Split provider selection and model rendering out of run_agent_step.go
run_agent_step.go is about spawning a container; who serves the model is a
different concern that had grown to ~380 lines inside it. agent_providers.go
now holds the whole seam: resolveJobProvider, the providerSetup registry,
bedrockSetup and its credential/discovery code, modelResolver, and
applyAgentModelEnv.

A file rather than a package: applyBedrockCreds reaches into utils.RunnerData
and the egress allowlist the spawn path owns, so a package boundary would mean
exporting a surface purely to keep the two halves talking — paying for
enforcement we do not need when the seam is already the interface.

Four AWS imports leave run_agent_step.go entirely, which is the useful
signal: the spawn path no longer refers to any provider's SDK.

Tests follow their code — bedrock_creds_test.go becomes agent_providers_test.go
and takes the resolveJobProvider tests with it.
